backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
bcrypt==4.0.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.3
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
executing==2.2.1
fastapi==0.136.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
Flask==3.1.2
fonttools==4.61.0
greenlet==3.4.0
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
ImageIO==2.37.3
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy-loader==0.5
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
packaging==25.0
pandas==2.1.3
parso==0.8.5
passlib==1.7.4
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2==2.9.12
psycopg2-binary==2.9.9
PuLP==2.9.0
pure_eval==0.2.3
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==1.0.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tornado==6.5.2
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
Werkzeug==3.1.4
wrapt==1.17.3
